# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PerceptionLab API",
    "version": "0.1.0",
    "description": "HTTP contract for PerceptionLab dataset, annotation, training, model, inference, export, and overlay workflows."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local development API"
    }
  ],
...
  "paths": {
...
    "/datasets/{dataset_id}/versions": {
      "post": {
        "operationId": "createDatasetVersion",
        "tags": ["Dataset Versions"],
        "summary": "Create an immutable dataset version.",
        "parameters": [{ "$ref": "#/components/parameters/DatasetId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateDatasetVersionRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created dataset version.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DatasetVersionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/RepositoryError" }
        }
      }
    },
...
    "/models": {
      "get": {
        "operationId": "listModels",
        "tags": ["Models"],
        "summary": "List registered models.",
        "responses": {
          "200": {
            "description": "Model list.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListModelsResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/RepositoryError" }
        }
      }
    },
    "/models/compare": {
      "post": {
        "operationId": "compareModels",
        "tags": ["Models"],
        "summary": "Compare registered models by a shared numeric metric.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompareModelsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model comparison.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelComparisonResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/RepositoryError" }
        }
      }
    },
    "/models/{model_id}": {
      "get": {
        "operationId": "getModel",
        "tags": ["Models"],
        "summary": "Read model detail.",
        "parameters": [{ "$ref": "#/components/parameters/ModelId" }],
        "responses": {
          "200": {
            "description": "Model detail.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/RepositoryError" }
        }
      }
    },
    "/models/{model_id}/promote": {
      "post": {
        "operationId": "promoteModel",
        "tags": ["Models"],
        "summary": "Promote a model and demote competing promoted models.",
        "parameters": [{ "$ref": "#/components/parameters/ModelId" }],
        "responses": {
          "200": {
            "description": "Promoted model.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/RepositoryError" }
        }
      }
    },
    "/models/{model_id}/infer": {
      "post": {
        "operationId": "runInference",
        "tags": ["Inference"],
        "summary": "Run inference on an image.",
        "parameters": [{ "$ref": "#/components/parameters/ModelId" }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["image"],
                "properties": {
                  "image": { "type": "string", "format": "binary" },
                  "confidence_threshold": {
                    "type": "number",
                    "format": "float",
                    "default": 0.25
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inference result.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InferenceResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaTypeError" },
          "500": { "$ref": "#/components/responses/RepositoryError" }
        }
      }
    },
    "/models/{model_id}/exports": {
      "post": {
        "operationId": "exportModel",
        "tags": ["Exports"],
        "summary": "Create a model export.",
        "parameters": [{ "$ref": "#/components/parameters/ModelId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateModelExportRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created model export.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ModelExportResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/RepositoryError" }
        }
      },
      "get": {
        "operationId": "listModelExports",
        "tags": ["Exports"],
        "summary": "List model exports.",
        "parameters": [{ "$ref": "#/components/parameters/ModelId" }],
        "responses": {
          "200": {
            "description": "Model exports.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListModelExportsResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/RepositoryError" }
        }
      }
    },
...
    "/inference-runs/{run_id}/overlay": {
      "post": {
        "operationId": "generateOverlay",
        "tags": ["Overlays"],
        "summary": "Generate an SVG overlay from stored inference detections.",
        "parameters": [{ "$ref": "#/components/parameters/InferenceRunId" }],
        "responses": {
          "201": {
            "description": "Generated overlay artifact.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OverlayResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/RepositoryError" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "DatasetId": {
        "name": "dataset_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
      "ModelId": {
        "name": "model_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "InferenceRunId": {
        "name": "run_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    },
    "responses": {
...
      "ValidationError": {
        "description": "Validation error.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "UnsupportedMediaTypeError": {
        "description": "Unsupported media type.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "RepositoryError": {
        "description": "Repository or adapter failure.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
...
      "CreateDatasetVersionRequest": {
        "type": "object",
        "required": ["version_name", "created_by"],
        "additionalProperties": false,
        "properties": {
          "version_name": { "type": "string" },
          "split_config": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          },
          "created_by": { "type": "string" }
        }
      },
      "DatasetVersionResponse": {
        "type": "object",
        "required": [
          "id",
          "dataset_id",
          "version_name",
          "sample_count",
          "annotation_count",
          "classes_snapshot",
          "split_config",
          "created_by"
        ],
        "additionalProperties": false,
        "properties": {
          "id": { "type": "string" },
          "dataset_id": { "type": "string" },
          "version_name": { "type": "string" },
          "sample_count": { "type": "integer", "minimum": 0 },
          "annotation_count": { "type": "integer", "minimum": 0 },
          "classes_snapshot": { "type": "array", "items": { "type": "string" } },
          "split_config": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          },
          "created_by": { "type": "string" }
        }
      },
...
      "ModelResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "training_job_id",
          "dataset_version_id",
          "model_family",
          "artifact_uri",
          "metrics_summary",
          "status"
        ],
        "additionalProperties": false,
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "version": { "type": "string" },
          "training_job_id": { "type": "string" },
          "dataset_version_id": { "type": "string" },
          "model_family": { "type": "string" },
          "artifact_uri": { "type": "string" },
          "metrics_summary": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          },
          "status": { "type": "string" }
        }
      },
      "ListModelsResponse": {
        "type": "object",
        "required": ["models"],
        "additionalProperties": false,
        "properties": {
          "models": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ModelResponse" }
          }
        }
      },
      "CompareModelsRequest": {
        "type": "object",
        "required": ["model_ids", "metric_name"],
        "additionalProperties": false,
        "properties": {
          "model_ids": {
            "type": "array",
            "minItems": 2,
            "items": { "type": "string" }
          },
          "metric_name": { "type": "string" }
        }
      },
      "ModelComparisonResponse": {
        "type": "object",
        "required": ["metric_name", "direction", "best_model_id", "models"],
        "additionalProperties": false,
        "properties": {
          "metric_name": { "type": "string" },
          "direction": {
            "type": "string",
            "enum": ["higher_is_better", "lower_is_better"]
          },
          "best_model_id": { "type": "string" },
          "models": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ModelComparisonEntryResponse" }
          }
        }
      },
      "ModelComparisonEntryResponse": {
        "type": "object",
        "required": [
          "rank",
          "model_id",
          "name",
          "version",
          "metric_value",
          "metrics_summary"
        ],
        "additionalProperties": false,
        "properties": {
          "rank": { "type": "integer", "minimum": 1 },
          "model_id": { "type": "string" },
          "name": { "type": "string" },
          "version": { "type": "string" },
          "metric_value": { "type": "number" },
          "metrics_summary": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          }
        }
      },
      "CreateModelExportRequest": {
        "type": "object",
        "required": ["format"],
        "additionalProperties": false,
        "properties": {
          "format": { "type": "string", "enum": ["onnx", "coreml"] }
        }
      },
      "ModelExportResponse": {
        "type": "object",
        "required": ["id", "model_id", "format", "artifact_uri", "status", "error_message"],
        "additionalProperties": false,
        "properties": {
          "id": { "type": "string" },
          "model_id": { "type": "string" },
          "format": { "type": "string" },
          "artifact_uri": { "type": ["string", "null"] },
          "status": { "type": "string" },
          "error_message": { "type": ["string", "null"] }
        }
      },
      "ListModelExportsResponse": {
        "type": "object",
        "required": ["exports"],
        "additionalProperties": false,
        "properties": {
          "exports": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ModelExportResponse" }
          }
        }
      },
      "InferenceResponse": {
        "type": "object",
        "required": ["run_id", "model_id", "latency_ms", "detections"],
        "additionalProperties": false,
        "properties": {
          "run_id": { "type": "string" },
          "model_id": { "type": "string" },
          "latency_ms": { "type": "integer", "minimum": 0 },
          "detections": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/DetectionResponse" }
          }
        }
      },
      "DetectionResponse": {
        "type": "object",
        "required": ["class_id", "class_name", "confidence", "bbox", "distance_m"],
        "additionalProperties": false,
        "properties": {
          "class_id": { "type": "integer", "minimum": 0 },
          "class_name": { "type": "string" },
          "confidence": { "type": "number", "format": "float", "minimum": 0, "maximum": 1 },
          "bbox": { "$ref": "#/components/schemas/DetectionBboxResponse" },
          "distance_m": { "type": ["number", "null"], "format": "float" }
        }
      },
      "DetectionBboxResponse": {
        "type": "object",
        "required": ["x", "y", "width", "height"],
        "additionalProperties": false,
        "properties": {
          "x": { "type": "number", "format": "float" },
          "y": { "type": "number", "format": "float" },
          "width": { "type": "number", "format": "float" },
          "height": { "type": "number", "format": "float" }
        }
      },
      "OverlayResponse": {
        "type": "object",
        "required": ["inference_run_id", "artifact_uri", "labels"],
        "additionalProperties": false,
        "properties": {
          "inference_run_id": { "type": "string" },
          "artifact_uri": { "type": "string" },
          "labels": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "additionalProperties": false,
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorBody" }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": ["code", "message"],
        "additionalProperties": false,
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" }
        }
      }
    }
  }
}